Moves database related functionality from transactions_csv_loader into transactions table manager. Updates test structure. Makes DuplicateError more abstract and moves to managers.common.
