+++ budgy/database_modules/managers/transaction_manager.py
@@ -15,29 +15,22 @@ Variables:
 """
 # Standard library imports
 from datetime import datetime
-from typing import Generator
+from typing import Generator, List
 import os
 
 # Third party imports
 import pandas as pd
 
 # Custom imports
-from pleasant_database import DatabaseFile, DatabaseManager
-
-
-class DuplicateError(Exception):
-    """Raised when a duplicate filepath is detected in the updates table."""
-    def __init__(self, filepath, table, message="Duplicate entry:"):
-        self.filepath = filepath
-        self.table = table
-        super().__init__(f"{message} {filepath}")
+from pleasant_database import DatabaseFile, DatabaseIntegrityError, DatabaseManager
 
 # Local imports
-from budgy.database_modules.managers.common import DB_FILE, convert_datetime_nums_to_range, format_column_names
-from budgy.database_modules.models.common import TableStatus
-from budgy.database_modules.models.transactions import TransactionsTable, UpdatesTable
+from budgy.database_modules.managers.common import DuplicateError, DB_FILE, convert_datetime_nums_to_range, format_column_names
+from budgy.database_modules.models.common import Field, TableStatus
+from budgy.database_modules.models.transactions import TransactionsTable, UpdatesTable, transaction_columns
 from budgy.utils.analysis_utils import PrimaryCategories, DetailedCategories, CategoriesEnum
 from budgy.utils.file_utils import EDirectories, LoggingExtras, get_csv_filenames
+from budgy.database_modules.io.transactions_csv_loader import iter_val_csv_file
 
 # initialize module logger
 import logging
@@ -127,6 +120,114 @@ class TransactionsTableManager(DatabaseManager):
 
     def __init__(self, db_file: DatabaseFile):
         super().__init__(TransactionsTable, db_file)
+
+
+    def _update_categories_if_diff(self, record: dict) -> None:
+        """
+        If a duplicate transaction is detected based on the base hash, check and update categories
+        if they differ from the existing database record.
+
+        Args:
+            record (dict): the record to check for duplicates and update categories for
+        """
+        base_hash = record[TransactionsTable.base_hash.name]
+        logger.debug(f"Checking for category difference between duplicates based on base hash: {base_hash}")
+
+        db_records = self.fetch_items_by_attribute(base_hash=base_hash)
+
+        for db_record in db_records:
+            if record[TransactionsTable.detailed_category.name] == db_record.detailed_category:
+                logger.debug(f"Categories are the same for record with base hash: {base_hash}. No update needed.", extra={LoggingExtras.BASE_HASH: base_hash})
+                continue
+            else:
+                try:
+                    self.update_item(
+                        item_id=db_record.id,
+                        primary_category=record[TransactionsTable.primary_category.name],
+                        detailed_category=record[TransactionsTable.detailed_category.name]
+                    )
+                except Exception as e:
+                    logger.exception(f"Exception encountered during category update for base hash: {base_hash}", extra={LoggingExtras.BASE_HASH: base_hash})
+                    raise e
+
+
+    def upload_csv(
+            self,
+            csv_filepath: str,
+            columns: List[Field] = transaction_columns,
+            updates_db_manager: 'UpdatesTableManager' = None
+        ) -> None:
+        """
+        Converts and validates the new transactions line by line then uploads to the transactions database.
+        Also enforces that no csv can be uploaded if it already has a posted upload with completed status.
+
+        Args:
+            csv_filepath (str): the filepath of the csv being uploaded
+            columns (List[Field]): the column mapping and conversion information for the csv upload
+            updates_db_manager (UpdatesTableManager): the database manager for the updates table
+        """
+        if updates_db_manager is None:
+            updates_db_manager = updates_manager
+
+        logger.info(f"Beginning upload of CSV file to database: {os.path.basename(csv_filepath)}", extra={LoggingExtras.FILE: csv_filepath})
+        logger.performance(f"Beginning csv upload process for {csv_filepath}", process_id=LoggingExtras.UPLOAD)
+
+        transactions_original_state = self.to_dataframe()
+
+        for record in iter_val_csv_file(csv_filepath, columns):
+
+            if record[TransactionsTable.base_hash.name] in transactions_original_state[TransactionsTable.base_hash.name].values:
+                self._update_categories_if_diff(record)
+
+            else:
+                try:
+                    self.add_item(**record)
+
+                except DatabaseIntegrityError:
+                    pass
+
+                except Exception as e:
+                    logger.exception(f"Exception encountered during data upload to {self}", extra={LoggingExtras.RECORD: record})
+                    updates_db_manager.generate_update_entry(csv_filepath, TableStatus.INCOMPLETE)
+                    raise e
+
+        logger.info(f"Completed upload of CSV file to database: {os.path.basename(csv_filepath)}", extra={LoggingExtras.FILE: csv_filepath})
+        updates_db_manager.generate_update_entry(csv_filepath, TableStatus.COMPLETE)
+        logger.performance(f"Completed csv upload process for {csv_filepath}", process_id=LoggingExtras.UPLOAD)
+
+
+    def upload_all_csvs(
+            self,
+            columns: List[Field] = transaction_columns,
+            updates_db_manager: 'UpdatesTableManager' = None,
+            csv_dir: str = EDirectories.CSV_DIR
+        ) -> None:
+        """
+        Iterates through all csv files in csv_dir and uploads only those not yet successfully uploaded.
+
+        Args:
+            columns (List[Field]): the column mapping and conversion information for the csv upload
+            updates_db_manager (UpdatesTableManager): the database manager for the updates table
+            csv_dir (str): path to the directory where the function should search for csv files to upload
+        """
+        if updates_db_manager is None:
+            updates_db_manager = updates_manager
+
+        logger.info(f"Beginning upload of all csv files in {csv_dir} to {self.table_name}")
+        failed_files = []
+
+        for csv_filepath in updates_db_manager.iter_csv_not_uploaded(csv_directory=csv_dir):
+            try:
+                self.upload_csv(csv_filepath, columns=columns, updates_db_manager=updates_db_manager)
+
+            except Exception:
+                logger.warning(f"Failed to upload {csv_filepath} to {self.table_name}", extra={LoggingExtras.FILE: csv_filepath})
+                failed_files.append(csv_filepath)
+
+        if failed_files:
+            logger.warning(f"Batch upload completed with {len(failed_files)} files failed")
+        else:
+            logger.info(f"New CSV file upload complete.")
 
 
     def retrieve_records_by_attribute_over_period(self, month: int=None, year: int=None, **kwargs) -> pd.DataFrame:
